requirements.txt
Django==5.1.2
djangorestframework==3.15.2
Pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
